Intensity exists
Intensity now exists in a barebones form. Will probably need more code,
and I need to consult with you on what you actually want intensity to
be. I also need to rotate the arrow, put it in the right spot, and size
it properly, as well as figure out how to make it invisible. But hey,
now it's a thing.